requirements.txt
numpy
opencv-python
torch
torchvision
phycv
scipy
Pillow
tqdm
diffusers
transformers
einops
accelerate